requirements.txt
Django==1.7.7
Markdown==2.6.1
django-filter==0.9.2
djangorestframework==3.1.1
mysqlclient==1.3.6
pytz==2015.2
requests==2.6.0